pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "metfish"
authors = [
  { name="Andrew Tritt", email="ajtritt@lbl.gov" },
  { name="Steph Prince", email="smprince@lbl.gov" },
  { name="Feng Yu", email="fyu2@lbl.gov" },
]
description = "A package for managing MetFish LDRD code"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: BSD License",
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
keywords = [
    "python",
]

dependencies = [
    "biopython",
    "pandas",
    "numpy",
    "scipy",
    "periodictable",
    "biopandas",
    "prody",
]

dynamic = ["version"]

[project.scripts]
calc-pr = "metfish.commands:get_Pr_cli"
extract-seq = "metfish.commands:extract_seq_cli"
generate-nma-conformers = "metfish.commands:generate_nma_conformers_cli"

[project.optional-dependencies]
training = ["wandb==0.16.6",
            "torch==1.12.1",
            "lightning==2.1.0",
            "pytorch-lightning==2.0.4",
            "openfold @ git+https://github.com/aqlaboratory/openfold.git@103d037",
            "ml-collections==0.1.0",
            "dm-tree==0.1.6",
            "lightning==2.1.0",
            "modelcif==0.7",
            "biopython==1.79",   # specific versions of main dependencies for reproducibility
            "pandas==1.5.3",
            "numpy==1.22.4",
            "scipy==1.7.2",
            "periodictable==2.0.2",
            "biopandas==0.5.1",
            ]

# for analysis and data visualization
viz = ["py3dmol==2.4.2",
       "seaborn==0.13.2",
       "matplotlib==3.8.4",
       "ipywidgets==8.1.5",
       "natsort==8.4.0"]

# for testing
dev = ["black==23.3.0",
       "codespell==2.2.4",
       "coverage==7.2.5",
       "pre-commit==3.3.1",
       "pytest==7.3.1",
       "pytest-cov==4.0.0",
       "ruff==0.0.265"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
